tests/wpt/tests/css/css-flexbox/parsing/order-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: getComputedStyle().order</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-order">
<meta name="assert" content="order computed value is the specified integer.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("order", "0");
test_computed_value("order", "123");
test_computed_value("order", "-45");
</script>
</body>
</html>
